+++ client-vue/src/components/Generic/Actions/PrinterDeleteAction.vue
@@ -8,7 +8,7 @@ import Vue from "vue";
 import FileControlList from "@/components/PrinterList/FileControlList.vue";
 import { Prop } from "vue-property-decorator";
 import { Printer } from "@/models/printers/printer.model";
-import { printersState } from "@/store/printers/printers";
+import { printersState } from "@/store/printers.state";
 
 @Component({
   components: { FileList: FileControlList }
