+++ client-vue/src/store/index.ts
@@ -3,6 +3,7 @@ import Vuex from "vuex";
 
 Vue.use(Vuex);
 
+// https://betterprogramming.pub/the-state-of-typed-vuex-the-cleanest-approach-2358ee05d230
 const store = new Vuex.Store({});
 
 export default store;
